requirements.txt
pytesseract
Pillow
opencv-python
numpy 